Finish initial Netlify CMS setup
This adds folders and initial index files for top-level content, development and
interaction design sections. The contents of the previous UX folder have been
moved into the interaction design folder

It also updates the admin config to include these three sections, and to use
file/folder paths with `src/` included. The latter is from a change to the
Playbook's admin config in commit 9eb5e7f, which fixed an issue where content was
not displayed or editable when opening a page for editing in the CMS (though this
was not reproducible locally)

@see https://github.com/dxw/playbook/pull/1001