Use assets/images for Netlify media folder
After re-enabling image uploads in 4c236aa, images were successfully
uploaded but they went into a project root `/_images` folder, which
appears not to be included in the Jekyll build. Links to images are not
working in production

This updates the media folder to `src/assets/images`. `src/assets` is
a more standard place to store images (and stylesheets), and files
placed there are already included in the build

This works locally without `src/` included, so the hope is that like for
`file` and `folder` Netlify config attributes, including it with `src/`
will work in production