Redesign navigation
This change reworks how navigation is displayed to a user in the side navigation - which should make it easier for a user to understand the structure of a whole site.

- A user can see related child pages for the page they are currently viewing,
- A user can always see the top level sections, and their relation to the child links for the section they are currently viewing
- The currently active link is underlined

Issues: this redesign works well for larger screens, but not when a user is viewing content on smaller screens (e.g. mobile)