Update nested menu to be collapsible
Instead of conditionally showing menu items if a user is on a section page or a page within a section, introducing collapsible nested menus using css only.

If a user is on a section page, or page within that section, that section of the menu will be expanded by default.

Add visually-hidden class - this is taken from https://github.com/dxw/playbook/blob/main/src/_sass/_accessibility.scss