Change toggle nav sections to use button
Previously we were using a form checkbox to toggle the expanded or collapsed state of the navigation section links - which meant that we did not need to use JavaScript.

However the trouble with using form elements is their longstanding association with the collection of data - and the use of a button with JavaScript would be more semantically correct.

Here, replacing the use of a checkbox with a button, and adding a toggle function which is called on clicking the button to manage the `aria-expanded` state.