Add a no-JavaScript fall-back
If a user does not have JavaScript they will not be able to control the buttons to toggle the navigation sections.

In this case, hiding the buttons (and arrow graphic). The links for a section will be shown when a vistor is on the section page, or one of the pages in that section, but the user will not be able to expand and collapse links in other sections.