Add expandable menu for mobile screens
This changes the way the menu is displayed on mobile screens. A 'hamburger' menu button is displayed in the mobile navigation which shows or hides the site navigation.

Currently using google material icon font, as shows an icon based on an appropriate keyword.

We will need a JavaScript fallback route, which could either be to show the menu as currently in the dxw playbook, or to use a checkbox toggle functionality rather than a button - as this does not rely on JavaScript.