Fix blank links appearing in nav for redirects
I have added some redirects using the redirect_from, but these links are included in the site pages list set of data, which is used to build the site navigation. As these redirects have no visible label, can easily skip a navigation item if the label is falsy.

There may be a better way to do this in our page filters.